backend/requirements.txt
Flask==3.0.3
pymongo==4.10.1
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Gunicorn
flask-cors
beautifulsoup4